move handleCategoryToggle from SideBarCategories to actions.js
